added android-wrapper
android-wrapper is a tool which allows you to run executables built
using the compilers shipped as part of the Android NDK on your linux
machine. It does that by using the android linker and generating a
customized configuration file for it.

Currenly only provides libraries for x86_64 and aarch64 as that is what
I'm targeting, but with a bit for tweaking android-wrapper and
produce-tree, you should be able to get the extract the libraries for
armv7 and x86 too and run those.

Running the linker on your linux machine requires that it has SELinux.