Added libicu, liblog and libz
This is due to the libararies shipped as part of the NDK being
"fauxlibs" that have every function call just return immediatly.

They'll still be included, but if I end up needing to call stuff from
the (such was the case for libicu, liblog and libz) they'll have to be
taken from the system image.